Fix space settings update
```
After the apply operation, the provider still indicated an unknown value for
mondoo_space.foo.space_settings.terminated_assets_configuration.
All values must be known after apply, so this is always a bug in the provider
and should be reported in the provider's own repository. Terraform will still
save the other known object values in the state.
```

Changed it to take unknown values from the plan